Rename ConfiguresOrchestra to OrchestratesService
